Add mock plugins and invoke them from manager
This change adds a few mock plugins to explore the design that is
evolving for this project. So far we have four types of plugins:
clouds, stores, checks, and alerts.

Here is a brief overview of how these plugins interact with each other:
Data is read from clouds and fed to stores and checks. Checks evaluate
the records that they are fed with and generate events when a check-rule
matches a record. These events are sent as alerts to destination stores
or recipients.