Add defaults for all MongoDBStore parameters
This change adds sensible default values for all parameters for the
`MongoDBStore` plugin. These default values would help new users as well
as developers who are trying out `MongoDBStore` on a local system to run
it with minimal configuration.

The defaults are chosen such that the plugin can connect to a locally
running MongoDB that requires no authentication. The database name is
set to `cloudmarker` by default. This is consistent with the default
path of `/tmp/cloudmarker` in `FileStore`.