Increase logging level of root logger to INFO
Prior to this change, the logging level of the root logger was set to
DEBUG. This causes debug logs like the following to appear in the logs:

    DEBUG adal-python:121 - b4cf52ce-d62b-418c-b581-402845596e76 -
    Authority:Performing instance discovery: ...
    DEBUG adal-python:121 - b4cf52ce-d62b-418c-b581-402845596e76 -
    Authority:Performing static instance discovery
    DEBUG adal-python:121 - b4cf52ce-d62b-418c-b581-402845596e76 -
    Authority:Authority validated via static instance discovery

DEBUG logs are not necessary during normal functioning of the project.
They are useful only while troubleshooting. Therefore, in this change,
the logging level for the root logger has been raised to INFO.