Search for user configuration in popular locations
Since Cloudmarker is supposed to serve as a CLI tool that one can run
with `cloudmarker -n` as well as a service that one may run as
`cloudmarker`, it is convenient if Cloudmarker searches certain popular
paths where configuration files usually reside. This would prevent the
users or service unit definition from having to specify configuration
files as arguments to the `-c` or `--config` option.

With this change, it now looks for a configuration file at the following
locations in the order specified:

  - /etc/cloudmarker.yaml
  - ~/.cloudmarker.yaml
  - ~/cloudmarker.yaml
  - cloudmarker.yaml (i.e., in the current working directory)

The base configuration file (which was named `config.base.yaml`) earlier
has been removed from the file system. There is no counterpart for it
now on the file system. Instead the base configuration is now built into
the code in a new module named `baseconfig`. This has been done for the
following reasons:

  - Provide reasonable defaults as built-in. This is consistent with how
    other tools work, e.g., Vim, Bash, etc. ship with defaults and we
    write `~/.vimrc`, `~/.bashrc`, etc. only when we want to override
    these defaults.

  - The users were never supposed to edit `config.base.yaml` anyway. By
    removing it from the file system, the users do not have to worry
    about notions like base configuration and user configuration. They
    do not have to be careful about not editing the base configuration.
    The users can still refer to the base configuration if they wish to
    by entering `cloudmarker --print-base-config`.

Since multiple configuration paths are involved, the logging has been
improved to show what configuration paths are being looked for and which
configuration paths are being found.