Do not log to file until fully configured
Prior to this change, we configure the logger with the full logger
configuration in `baseconfig.py` as soon as Cloudmarker starts. This is
a problem because it causes Cloudmarker to attempt logging to a file
path that may not exist.

Further, the default path is `logs/cloudmarker.log` but it is possible
that a user does not want any log files to be written under the current
directory. This is particularly a problem when, for example, Cloudmarker
is installed under `/usr` because this path is defined to be read-only
by [FHS 2.3].

The user cannot change the path of the log file for the first few log
messages via a configuration file, because the first few logs are
written even before the configuration files have been read. Therefore,
with this change, we ensure that we begin logging to a file only after
all configuration files have been read, so that the user has a chance to
specify which path to write the logs to.

To accomplish this change, we use the logger configuration in
`baseconfig.py` but we initially override the logger handlers to log to
console only, i.e., avoid logging to file.

[FHS 2.3]: http://refspecs.linuxfoundation.org/FHS_2.3/fhs-2.3.html