Put all plugin configs under a key named 'plugins'
Prior to this change, the various plugins were configured under separate
keys named `clouds`, `stores`, `events`, and `alerts`. This led to
duplicate plugin configs when the same store plugin is used as alert
plugin as well.

For example, in order to use `FileStore` and `SplunkHECStore` plugins
for both storing cloud records and event records, we need to duplicate
the configuration for them twice, once under `stores` and again under
`alerts`:

    stores:
      filestore:
        plugin: cloudmarker.stores.filestore.FileStore

      splunkstore:
        plugin: cloudmarker.stores.splunkhecstore.SplunkHECStore
        params:
          uri: https://localhost:8088/services/collector
          token: token123
          index_name: main
          ca_cert: false

    alerts:
      filestore:
        plugin: cloudmarker.stores.filestore.FileStore

      splunkstore:
        plugin: cloudmarker.stores.splunkhecstore.SplunkHECStore
        params:
          uri: https://localhost:8088/services/collector
          token: token123
          index_name: main
          ca_cert: false

    audits:
      mockaudit:
        stores:
          - filestore
          - splunkstore
        alerts:
          - filestore
          - splunkstore

With this change, all plugin configuration is placed under a single key
named `plugins` as shown in the following example:

    plugins:
      filestore:
        plugin: cloudmarker.stores.filestore.FileStore

      splunkstore:
        plugin: cloudmarker.stores.splunkhecstore.SplunkHECStore
        params:
          uri: https://localhost:8088/services/collector
          token: token123
          index_name: main
          ca_cert: false

    audits:
      mockaudit:
        stores:
          - filestore
          - splunkstore
        alerts:
          - filestore
          - splunkstore

Thus, there is no need to configure the same plugin twice as a store and
as an alert. If the same plugin can be used as both a store plugin and
as an alert plugin, it can be defined only once and reused as store as
well as alert.