Handle empty destination_port_range gracefully
See previous commit message for an example of an example of an Azure
security rule that contains both `destination_port_range` and
`destination_port_ranges` fields in the same record with the value of
the former as an empty string.

Since an empty string for this field indicates a missing port range,
this change ensures that we ignore it.