Replace os.rename() with os.replace() in FileStore
The cross-platform behavior of `os.rename()` is inconsistent.
While trying to setup a development environment on Windows,
`os.rename()` caused errors to be thrown as seen below.

    File "c:\users\...", line 89, in done
        os.rename(tmp_file_path, json_file_path)
    FileExistsError: [WinError 183] Cannot create a file when that file
    already exists: '/tmp/cloudmarker\\mymockaudit_mymockcloud.tmp' ->
    '/tmp/cloudmarker\\mymockaudit_mymockcloud.json'
    mymockaudit_filestore: Failed; done() error: FileExistsError:
    [WinError 183] Cannot create a file when that file already exists:
    '/tmp/cloudmarker\\no_worker.tmp' ->
    '/tmp/cloudmarker\\no_worker.json'

Changing `os.rename()` to `os.replace()` has fixed the issue, as
`os.replace()` always guarantees to replace the file regardless
of the platform.