Add baseconfig key for RDBMSEnforceTLSEvent
The `baseconfig.py` module contains built-in config keys that can be
referenced directly in the user's config file. In other words, a user
can use these config keys without having to define them.

This change adds an `rdbmsenforcetlsevent` config key to configure the
`RDBMSEnforceTLSEvent` plugin. This plugin is a good candidate for a
built-in base config key because this plugin accepts no parameters, so
once configured in hte base config, it can be used in any user config.