Identify recoverable Azure Key Vaults
There are two Azure Key Vault attributes which play an important role in
safeguarding Azure Key Vault. These are:

1. `enableSoftDelete`: Setting this Azure Key Vault  parameter to true
ensures that the KV and its associated objects remain recoverable for
the next 90 days. During the 90 day window the Key Vault can either be
recovered or purged (permanently deleted). After the end of 90 days
period, the Key Vault is automatically deleted forever.

2. `enablePurgeProtection`: Turning on this KV parameter by setting its
value to true makes sure that a Key Vault  can never be purged. This is
used in conjugation with the enableSoftDelete attribute. As
`enableSoftDelete` only ensures that the Key Vault is recoverable 90
days post delete, turning on purge protection ensures that the Key Vault
is not purged post delete. Once turned on, purge protection can not be
turned-off.

When an Azure KV is provisioned both these features are disabled. Azure
admins must make sure these two features are turned on for all the Key
Vaults  to protect them from any any accidental or malicious deletion.
There is no way to turn these two features on from the Azure portal.
Azure CLI or Azure PoweShell can be used to enable them.