Freeze versions of dependencies in dev environment
The dependencies have changed a lot since we began this project and as
a result, some of the unittests fail with the latest versions of some
of the Azure and GCP dependencies. This change freezes the
dependencies in development environment to the last known versions
with which the tests passed successfully.

In future, if we can afford the time, ideally we should upgrade all
dependencies to the most recent version possible and update the tests
so that they pass with the most recent versions of dependencies.