Print output of pip3 freeze in make deps target
The output of pip3 freeze would help us to get a pip3-readable log of
the entire set of all dependencies installed in Travis CI environment.
If we need to refer back to it in future, we can visit a build job
page on Travis CI and use this output.