Remove Python 3.5 from Travis CI build matrix
We need to upgrade PyYAML from 5.1 to 5.4 in order to resolve a few
arbitrary code execution vulnerabilities. However version 5.4 of
PyYAML is not supported for Python 3.5. See the `python_requires`
argument in `setup.py` of PyYAML source code at
https://github.com/yaml/pyyaml/blob/5.4/setup.py#L295 to confirm this.

Further, Python 3.5 reached end-of-life in September 2020, so it is
okay for this project to not support Python 3.5 anymore. We will no
longer test this project with Python 3.5.